Have constructing relative Name with length() == MAXNAME fail
The Name() constructor will currently fail if it's length() would
end up being longer than MAXNAME (255). However, there is an edge
case where a relative name can be exactly MAXNAME long. This is
legal but the lookup mechanism will then fail to make the name
absolute by appending an empty label to be able to use it for
lookups.