Make Record serializable again
Instead of relying on the unstable default Java serialization of each
record type, this uses a proxy object. The proxy object only contains
the DNS wire data of a record, which is guaranteed to be stable. This
avoids all concerns about compatibility when Record classes evolve and
with inheritance.

Closes #132
See #114