Fix Name constructor failing with max length relative name and root origin
The Name constructor failed when passing a max-length (253 chars) relative name and Name.root as origin, while this name is technically allowed.

The failure was caused by the Name constructor having the local "absolute" variable wrongly set to false in this case. This change fixes this by updating this flag when the origin is present.

Previous to this fix the workaround was to pass a null origin and manually append the final dot (root) to the string: in this case the constructor would work (as demonstrated by the "ctor_max_length_abs" test.

With this change I've also added a few tests for these edge-cases.

(cherry picked from commit 84e0e2365ba533fbb247145bc44991ea4b401782)

Closes #289